requirements.txt
streamlit
google-api-python-client
google-generativeai
python-dotenv
groq